requirements.txt
uvicorn
fastapi
pandas
GitPython
